# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Microsoft Device Code Auth API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /code-status/{userCode}:
    get:
      operationId: getCodeStatus
      tags: [auth]
      summary: Get status of a device code
      parameters:
        - name: userCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Code status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeStatusResponse"
        "404":
          description: Code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    CodeStatusResponse:
      type: object
      properties:
        user_code:
          type: string
        status:
          type: string
          enum: [POLLING, SUCCESS, EXPIRED]
        app:
          type: string
          description: Display name of the app
        alias:
          type: string
          description: Short alias used to generate this code
        generated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        expires_in:
          type: integer
        last_polled_at:
          type: string
          format: date-time
          nullable: true
      required:
        - user_code
        - status
        - app
        - alias
        - generated_at
        - expires_at
        - expires_in
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...